requirements.txt
Flask==3.1.2
openai==1.109.1
python-dotenv==1.1.1
flask-cors==5.0.1
gunicorn==23.0.0